feat>(basics-cpucaches): cache placement policy and false sharing
False sharing demonstration to show how using padding in a struct
can improve CPU cachine and therefore improve application performance
especially when using goroutines to achieve concurrency.

Benchmark tests have been added to show the difference between the two
approaches. Note that using padding is one of the solutions and instead
using channels to communicate the result between the goroutines is an
alternative solution as well.