feat(basics-cpu-caches): instruction level parallelism
This shows instruction level parallelsm (ILP) in Go and
how that can affect performance. This is especially the case with
CPU caches and how the CPU can parallelize instructions to improve
performance.

One function shows data hazards when reading and incrementing a value
in an array, while the other shows how this can be bypassed to enable
the CPU to handle ILP by assigning that value to a new variable and
using
that variable instead.

The second method shows an improvement boost over the first
as it allows the CPU to utilize ILP.